Various standard classes and move logic from build.pl
The main script was getting far too long and hard to cleanly navigate, so split it into a frontend main, and a backend, AltimatOS::Build class.

In addition to moving code into a subclass, classes were created to make modern takes on older behaviour in Perl that normally don't have a uniform exception pattern.

Classes:
- Console::IO: A perl-ish implementation of cout
- File::IO: file IO with exceptions
- File::IO::JSON: JSON5-based IO with exceptions
- GNUzip::Decompress: gunzip with exceptions
- Sys::Error:

Signed-off-by: Gary Greene <greeneg@tolharadys.net>